@@ -14,19 +14,22 @@
     "clean": "rimraf dist/ .astro/"
   },
   "dependencies": {
-    "@astrojs/check": "^0.7.0",
-    "@astrojs/starlight": "^0.24.3",
-    "astro": "^4.10.3",
-    "sharp": "^0.33.4",
-    "typescript": "^5.4.5"
+    "@astrojs/check": "^0.9.3",
+    "@astrojs/starlight": "^0.27.1",
+    "astro": "^4.15.4",
+    "sharp": "^0.33.5",
+    "typescript": "^5.5.4"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
+    "eslint": "^9.10.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-astro": "^1.2.2",
-    "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-prettier": "^5.1.3",
-    "rimraf": "^5.0.7"
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-astro": "^1.2.3",
+    "eslint-plugin-import-x": "^4.2.1",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-react-hooks": "5.1.0-rc-fb9a90fa48-20240614",
+    "globals": "^15.9.0",
+    "rimraf": "^6.0.1",
+    "typescript-eslint": "^8.4.0"
   }
 }
